11. Microservices Concepts/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.1
drf-yasg==1.21.10
pytz==2025.2
PyYAML==6.0.2
uritemplate==4.2.0
inflection==0.5.1
packaging==25.0
